Update python/cudf/cudf/core/series.py
empty case returns same type as self

+++ python/cudf/cudf/core/series.py
@@ -144,7 +144,7 @@ class Series(object):
             raise TypeError("Data must be numeric")
 
         if len(self) == 0:
-            return cupy.asarray([])
+            return cupy.asarray([], dtype=self.dtype)
 
         return cupy.asarray(self._column.to_gpu_array())
